Add: ResponseEntity들을 구현하는 클래스들 생성
